fixed more warning messages of -Wsign-compare type
Import internal commit a9a9707739204aca5762b1e3d30ecd2876f202d8

+++ groups/bsl/bslstl/bslstl_unorderedset.t.cpp
@@ -2489,13 +2489,13 @@ void TestDriver<KEY, HASH, EQUAL, ALLOC>::testCase22()
 
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH <= da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH <= (size_t) da.numBlocksInUse());
 
             ObjStlAlloc mY(X);  const ObjStlAlloc& Y = mY;
 
             ASSERTV(LINE, 0 == verifyContainer(Y, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    2 * TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    2 * TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
 
             ObjStlAlloc mZ;  const ObjStlAlloc& Z = mZ;
 
@@ -2503,7 +2503,7 @@ void TestDriver<KEY, HASH, EQUAL, ALLOC>::testCase22()
 
             ASSERTV(LINE, 0 == verifyContainer(Z, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    2 * TYPE_ALLOC * LENGTH <= da.numBlocksInUse());
+                    2 * TYPE_ALLOC * LENGTH <= (size_t) da.numBlocksInUse());
         }
 
         CONT.resetIterators();
@@ -2513,7 +2513,7 @@ void TestDriver<KEY, HASH, EQUAL, ALLOC>::testCase22()
             mX.insert(BEGIN, END);
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
         }
 
         CONT.resetIterators();
@@ -2528,7 +2528,7 @@ void TestDriver<KEY, HASH, EQUAL, ALLOC>::testCase22()
             }
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH <= da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH <= (size_t) da.numBlocksInUse());
         }
 
         ASSERTV(LINE, da.numBlocksInUse(), 0 == da.numBlocksInUse());
@@ -4985,7 +4985,7 @@ void TestDriver<KEY, HASH, EQUAL, ALLOC>::testCase7()
 
                 Obj Y11(X, &oa);
 
-                ASSERT(0 == LENGTH || oa.numBlocksTotal() > A);
+                ASSERT(0 == LENGTH || (size_t) oa.numBlocksTotal() > A);
 
                 // Due of pooling of memory alloctioon, we can't predict
                 // whether this insert will allocate or not.
